+++ FWCore/Framework/bin/BuildFile.xml
@@ -70,7 +70,7 @@
   <use   name="FWCore/ParameterSet"/>
   <use   name="FWCore/ParameterSetReader"/>
 </bin>
-<bin   name="cmsRunPB11" file="cmsRunPB11.cpp">
+<bin   name="cmsRunPyDev" file="cmsRunPyDev.cpp">
   <use   name="tbb"/>
   <use   name="boost"/>
   <use   name="boost_program_options"/>
@@ -84,5 +84,5 @@
   <use   name="FWCore/ServiceRegistry"/>
   <use   name="FWCore/Utilities"/>
   <use   name="FWCore/ParameterSet"/>
-  <use   name="FWCore/PyBind11ParameterSet"/>
+  <use   name="FWCore/PyDevParameterSet"/>
 </bin>
