+++ FWCore/PyDevParameterSet/src/MakePyBind11ParameterSets.cc
@@ -1,8 +1,8 @@
-#include "FWCore/PyBind11ParameterSet/interface/MakePyBind11ParameterSets.h"
+#include "FWCore/PyDevParameterSet/interface/MakePyBind11ParameterSets.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 
-#include "FWCore/PyBind11ParameterSet/interface/Python11ParameterSet.h"
-#include "FWCore/PyBind11ParameterSet/interface/PyBind11ProcessDesc.h"
+#include "FWCore/PyDevParameterSet/interface/Python11ParameterSet.h"
+#include "FWCore/PyDevParameterSet/interface/PyBind11ProcessDesc.h"
 #include "initializePyBind11Module.h"
 #include <pybind11/embed.h>
