+++ FWCore/PyDevParameterSet/test/processbuilder_t.cppunit.cpp
@@ -9,7 +9,7 @@
 #include <cppunit/extensions/HelperMacros.h>
 
 #include <FWCore/ParameterSet/interface/ParameterSet.h>
-#include <FWCore/PyBind11ParameterSet/interface/PyBind11ProcessDesc.h>
+#include <FWCore/PyDevParameterSet/interface/PyBind11ProcessDesc.h>
 #include "FWCore/Utilities/interface/EDMException.h"
 
 #include <memory>
